Cleanup some circular imports in group
Breakout the registry into registry.py

Breakout the base entities into entity.py

This are no code changes other than splitting
into modules
