+++ homeassistant/components/group/const.py
@@ -4,3 +4,16 @@ CONF_HIDE_MEMBERS = "hide_members"
 CONF_IGNORE_NON_NUMERIC = "ignore_non_numeric"
 
 DOMAIN = "group"
+
+REG_KEY = f"{DOMAIN}_registry"
+
+GROUP_ORDER = "group_order"
+
+
+ATTR_ADD_ENTITIES = "add_entities"
+ATTR_REMOVE_ENTITIES = "remove_entities"
+ATTR_AUTO = "auto"
+ATTR_ENTITIES = "entities"
+ATTR_OBJECT_ID = "object_id"
+ATTR_ORDER = "order"
+ATTR_ALL = "all"
