+++ homeassistant/components/group/cover.py
@@ -43,7 +43,7 @@ from homeassistant.helpers import config_validation as cv, entity_registry as er
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import GroupEntity
+from .entity import GroupEntity
 from .util import reduce_attribute
 
 KEY_OPEN_CLOSE = "open_close"
